Rework 2019 day 13 part 2 visualizer
Now uses non-ASCII characters for display, and displays score on the
top.

It no longer uses a HashMap for the screen, instead executing draw
instructions directly as needed, and storing the x positions of
the ball and paddle only.

It also moves the cursor to the end of the screen at the end of execution.