2020 day 20
Got very caught up in efficiently storing only the edges for part 1,
couldn't figure out how to go from a packed bit representation to an
actual arrangement, and figure out all of the possible dimensions for a
given number of tiles, to allocate a 2d array of the needed size,
passing the data between C and Rust via FFI. I realized I was
overengineering and hyperoptimizing for a small part of part 1, and
couldn't even imagine how I'd actually check if the pieces fit together
or not, and would need to throw out most of my work if the inner 8x8
pixels was needed for part 2. (Spoiler: it was).

A few days later, I was drifting off to sleep, and suddenly realized
that I don't need to pre-allocate the grid, and if each edge has at most
one edge that matches a different piece, I can just start with 1 piece,
and cycle through pieces 1 by 1 to see if they fit any edges in
already-positioned pieces, then just double-check that the position is
empty and it fits with all neighbors.

The next day, I decided not to bother with the hyper-optimized,
bit-packed representation, and to just cycle through a queue of tiles,
removing them as they can be fit in, with a hashmap of coordinates to
pieces.