2020 day 23
My part 1 solution does everything from parsing the input, to simulating
the crab shuffle, to stringifying the output, entirely at compile time
within const blocks - the only non-const operation is actually printing
the output.

My initial part 2 solution took around 2 hours to run, but got the right
answer in the end. I had an idea for a different data structure that
seemed likely to be more efficient, but didn't work well with Rust's
semantics, and was outside of my comfort zone. I spent around 10 hours
figuring out how to implement it - it's a strange hybrid of a linked list
and an array in C, and got a part 2 solution built on that which runs in
around 0.3 seconds on my system.