2020 day 25
I have both a fully-const version that takes around 12-14 seconds for my
computer to compile and run with the `aoc-rustc` wrapper in `tooling`,
and a version that saves the actual calculation of the answer for
runtime, and can compile and run in around 110 to 125 ms.

I guess the compile-time const evaluator is much slower than running
native code, for more complex calculations.

The const version itself runs in around 1ms, while the version that
calculates at compile time runs in around 77 - 80 ms - not long at all,
but still noticable slower.

(All timings are based on running Bash's `time` builtin a few times
until I had a sense of the range of timings, and are by no means meant
to be scientific-quality data).