2021 day 21
I saw the short input, and got worried - short inputs usually mean big
challanges. Part 1 seemed quite simple, so I was scared of part 2. When
I saw the description, I saw that hundreds of trillions of games would
run with the sample input, so I figured I'd try memoization, but was
worried about running out of memory with that approach.

Nonetheless, I figured I'd give part 2 a go with a simple HashMap as a
cache, and gave it a try, with an eye on memory usage and fingers ready
to Ctrl+C the process. It didn't work, as I was adding the roll value,
rather than the space, to the score, but also did not use too much
memory. Fixing the mistake with the scoring, it runs in around 10-12
milliseconds on my system, and only allocates around 1.6 MiB (per
valgrind).