2021 day 22
Whew...

For part 1, I did it the more-or-less brute force way. The only clever
part of the approach was that I realized that working through the steps
backwards, the first step that would apply to a given cube would be the
one to go with, as it would override any prevous step, so I went through
all 101^3 cubes in the region, one by one, working on that, knowing full
well that the approach would not work at all for part 2.

For part 2, I had trouble trying to split cubes at intersections, so I
just created a list of every x coordinate that appeared, every y
coordinate that appeared, and every z coordinate that appeared, and
checked each individual region that created. This was rather slow, but
fast enough that it would take only a minute or two, not hours or worse,
so I decided to try multi-threading to speed it up. Probably due to my
inexperience in multithreading, I originally set it up so that each
thread would lock the mutex just long enough to take 1 item to check,
then work on it. The result was a system that slowed down with more
threads. My guess was that the threads were each waiting on the Mutex
lock more often than they were working, so I came up with a simple
wrapper iterator that bunches up the returned items into collections up
to a certain size, and returns those. Chunking tasks into Vecs of up to
1024, the time was brought down to a surprisingly consistent 10.89
seconds on my system.