2021 day 23
Whew. This was a tough one.

I had a nearly full-time 3-week gig start a few days after I reached
this one, and spent those few days reworking my drop-in generic
Dijkstra's algorithm implementation, while struggling to figure out a
good way to model the problem.

When the program was over, I came back with fresh eyes and came up with
a way to model part 1, albeit messily. I did miss some of the valid
states and messed up distance calculation, but got it working once I'd
addressed those issues.

For part 2, I switched to a cleaner daata model and implementation, and
got it working. Based on unscientific measurements using bash's "time"
builtin, it seems to be around 9 times slower than part 1, which seems
pretty reasonable given the added complexity of part 2.

Reworking the part 2 solution to solve part 1 appears to save a few
milliseconds, but it's a small enough difference that it may be due to
normal variation in run time.