2022 day 11
Part 1 was easy - it worked flawlessly second try, and the first try
only failed because I was splitting the input on "\n" instead of "\n\n".

The wording of part 2 made me believe it may go out of bounds of sized
integers. I saw that the example answer would need to be at least 32
bits in length.

Before even updating the loop to run 10000 times, I'd already pasted in
a least-common-multiple function, and I'd set it up to use that to
ensure that the worry factor could be kept under control.