2025 day 09
Part 2 was the first 2025 puzzle that provided a serious challenge to
me, though I knew (broadly) what I needed to do from the start. I was
having a shocking amount of trouble determining if a segment overlapped
a rectangle, and accounting for the fact that if and only if it's at the
edge of the rectangle, it's allowed.