2025 day 10
For part 1, I figured I'd use Dijkstra's algorithm, but in C this time.

I'd wanted to learn Z3 for a previous problem I was stuck on (2021 day
24), but hadn't taken the time. I was able to figure out how to use it
for part 2.

I was going to at least reuse the parsing code from the part 1 C
solution, but the C API is underdocumented, and it's not clear how to go
to or from the s-expression representation.

This is the first day that I used either 3rd-party libraries or
non-portable C extensions within the actual solution code, and had that
distinction even before I realized that trying to learn z3 was almost
required for part 2.