2025 day 11
This was a nice break after the last 2 days

Part 1, I used a `u64` just in case it was needed, but it turned out to
fit within a `u16`, so I switched to that right before part 2, but part
2 actually did need `u64`s to avoid overflow. That's a shame - if the
ResolvedRouteCount struct would fit in a 64-bit register, then that
might allow for some optimization.

That said, Bash's `time` builtin reporting a consistent "0m0.002s" for
almost all runs of either part, that optimization is entirely unneeded.