2025 day 12
I was trying to figure out what algorithm could possibly work for this
problem, and went to the subreddit for hints, and saw people talking
about the actual (non-sample) input making the problem much easier than
it would normally be, so I decided to see if I'd get the right answer by
just assuming each present takes a full 3x3 area, and counting the areas
large enough to fit 9 times the number of presents. That got the right
answer.